Francesco: added more fields for graphs and data extraction.  Next step is the implementation of a smart hashmap for speedup of the computations.

+++ openrocket/swing/src/net/sf/openrocket/gui/main/SimulationPanel.java
@@ -242,13 +242,13 @@ public class SimulationPanel extends JPanel {
 					return;
 				}
 
-				int trainMinutes = 1;
+				double trainMinutes = 1.0;
 				JPanel repPanel = new JPanel(new MigLayout());
-				String trainMinutesFromPanel = JOptionPane.showInputDialog(SimulationPanel.this, new Object[] {"How many minutes?", "", repPanel}, 1);
+				String trainMinutesFromPanel = JOptionPane.showInputDialog(SimulationPanel.this, new Object[] {"How many minutes?", "", repPanel}, 1.0);
 				try {
 					if (trainMinutesFromPanel == null)
 						return;
-					trainMinutes = Integer.parseInt(trainMinutesFromPanel);
+					trainMinutes = Double.parseDouble(trainMinutesFromPanel);
 				} catch (Exception exc) {
 					return;
 				}
